#6 [ui] mypage 화면 구성
